### 2021-11-22 #3943 @whedon
[PRE REVIEW]: PWSpy: A Python library dedicated to the analysis of Partial Wave Spectroscopic Microscopy data.
**Submitting author:** @nanthony21 (<a href="http://orcid.org/0000-0003-2882-2471">Nicholas Anthony</a>)
**Repository:** <a href="https://github.com/BackmanLab/PWSpy" target ="_blank">https://github.com/BackmanLab/PWSpy</a>
**Version:** 0.2.11
