### 2021-11-22 #3943 @nanthony21
The paper builds fine through [Github Actions](https://github.com/BackmanLab/PWSpy/actions/runs/1490905003), Docker, and the online previewer, I'm not sure what the issue is here.
