### 2021-11-22 3b87027
Add index to affiliation
