### 2021-11-22 #3943 @nanthony21
> @openjournals/dev can you help troubleshoot this error ☝️

It looks like it may have been because my `affiliations` section was missing `index: 1`. It's fixed now.
