### 2021-11-24 ea87a2e
Add utility class for running analyiss in parallel
