### 2021-11-29 #3957 @whedon
[REVIEW]: PWSpy: A Python library dedicated to the analysis of Partial Wave Spectroscopic Microscopy data.
**Submitting author:** @nanthony21 (<a href="http://orcid.org/0000-0003-2882-2471">Nicholas Anthony</a>)
**Repository:** https://github.com/BackmanLab/PWSpy
**Branch with paper.md** (empty if default branch):
**Version:** 1.0.0

- [x] #4072
