### 2021-11-29 3b6c210
ROI compilers depend just on an ROI rather than the ROIFile.
