### 2021-11-29 e86ef1d
Allow material index to be calculated from cauchy equation and add Glycerol using this method.
