### 2021-12-01 #3957 @jgostick
Hi @nanthony21, I have taken a look at your paper and it has a few shortcomings that you might want to address before the reviewers do.  It's a bit on the short side...I know that the joss paper is not meant to be a full length snooze-fest, but this is an opportunity for you to give an outline of how the code works, like which modules are offered, and maybe a few figures highlighting the output.  Also the statement of need really focuses on the ability to handle different file formats.  Could you not highlight the need in a more 'big picture' way, like consistent analyses between researchers, unit tested results for ensure accuracy, etc?
