### 2021-12-02 9c2659c
add a new example. Add licenses
