+++ examples/limitedOPDSigma.py
@@ -1,5 +1,22 @@
 
 # -*- coding: utf-8 -*-
+# Copyright 2018-2021 Nick Anthony, Backman Biophotonics Lab, Northwestern University
+#
+# This file is part of PWSpy.
+#
+# PWSpy is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# PWSpy is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with PWSpy.  If not, see <https://www.gnu.org/licenses/>.
+
 """
 This script is based on a matlab script written by Lusik Cherkezyan for NC.
 Nano uses this method to extract rms from phantom make from ChromEM cells embedded in resin.
