+++ src/pwspy/dataTypes/_data.py
@@ -891,8 +891,8 @@ class PwsCube(ICRawBase):
 
     _hdfTypeName = "ImCube"  # This is used for saving/loading from HDF. Important not to change it or old files will stop working.
 
-    def __init__(self, data, metadata: pwsdtmd.ICMetaData, processingStatus: ICRawBase.ProcessingStatus = None, dtype=np.float32):
-        assert isinstance(metadata, pwsdtmd.ICMetaData)
+    def __init__(self, data, metadata: pwsdtmd.PwsMetaData, processingStatus: ICRawBase.ProcessingStatus = None, dtype=np.float32):
+        assert isinstance(metadata, pwsdtmd.PwsMetaData)
         super().__init__(data, metadata.wavelengths, metadata, processingStatus=processingStatus, dtype=dtype)
 
     @property
@@ -903,7 +903,7 @@ class PwsCube(ICRawBase):
         return self.index
 
     @classmethod
-    def loadAny(cls, directory: str, metadata: pwsdtmd.ICMetaData = None,  lock: mp.Lock = None):
+    def loadAny(cls, directory: str, metadata: pwsdtmd.PwsMetaData = None, lock: mp.Lock = None):
         """
         Attempt to load a `PwsCube` for any format of file in `directory`
 
@@ -927,7 +927,7 @@ class PwsCube(ICRawBase):
                     raise OSError(f"Could not find a valid PWS image cube file at {directory}.")
 
     @classmethod
-    def fromOldPWS(cls, directory: str, metadata: pwsdtmd.ICMetaData = None,  lock: mp.Lock = None):
+    def fromOldPWS(cls, directory: str, metadata: pwsdtmd.PwsMetaData = None, lock: mp.Lock = None):
         """
         Loads from the file format that was saved by the all-matlab version of the Basis acquisition code.
         Data was saved in raw binary to a file called `image_cube`. Some metadata was saved to .mat files called
@@ -945,7 +945,7 @@ class PwsCube(ICRawBase):
             lock.acquire()
         try:
             if metadata is None:
-                metadata = pwsdtmd.ICMetaData.fromOldPWS(directory)
+                metadata = pwsdtmd.PwsMetaData.fromOldPWS(directory)
             with open(os.path.join(directory, 'image_cube'), 'rb') as f:
                 data = np.frombuffer(f.read(), dtype=np.uint16)
             data = data.reshape((metadata.dict['imgHeight'], metadata.dict['imgWidth'], len(metadata.wavelengths)), order='F')
@@ -956,7 +956,7 @@ class PwsCube(ICRawBase):
         return cls(data, metadata)
 
     @classmethod
-    def fromTiff(cls, directory, metadata: pwsdtmd.ICMetaData = None,  lock: mp.Lock = None):
+    def fromTiff(cls, directory, metadata: pwsdtmd.PwsMetaData = None, lock: mp.Lock = None):
         """
         Loads from a 3D tiff file named `pws.tif`, or in some older data `MMStack.ome.tif`. Metadata can be stored in
         the tags of the tiff file but if there is a pwsmetadata.json file found then this is preferred.
@@ -975,7 +975,7 @@ class PwsCube(ICRawBase):
             lock.acquire()
         try:
             if metadata is None:
-                metadata = pwsdtmd.ICMetaData.fromTiff(directory)
+                metadata = pwsdtmd.PwsMetaData.fromTiff(directory)
             if os.path.exists(os.path.join(directory, 'MMStack.ome.tif')):
                 path = os.path.join(directory, 'MMStack.ome.tif')
[…]
