+++ src/pwspy/dataTypes/_metadata.py
@@ -574,7 +574,7 @@ class FluorMetaData(MetaDataBase):
             return f.asarray()
 
 
-class ICMetaData(MetaDataBase, AnalysisManager):
+class PwsMetaData(MetaDataBase, AnalysisManager):
     """A class that represents the metadata of a PWS acquisition.
 
     Args:
@@ -591,14 +591,14 @@ class ICMetaData(MetaDataBase, AnalysisManager):
         from pwspy.analysis.pws import PWSAnalysisResults
         return PWSAnalysisResults
 
-    _jsonSchemaPath = os.path.join(_jsonSchemasPath, 'ICMetaData.json')
+    _jsonSchemaPath = os.path.join(_jsonSchemasPath, 'PwsMetaData.json')
     with open(_jsonSchemaPath) as f:
         _jsonSchema = json.load(f)
 
-    def __init__(self, metadata: dict, filePath: t_.Optional[str] = None, fileFormat: ICMetaData.FileFormats = None, acquisitionDirectory: t_.Optional[Acquisition] = None):
+    def __init__(self, metadata: dict, filePath: t_.Optional[str] = None, fileFormat: PwsMetaData.FileFormats = None, acquisitionDirectory: t_.Optional[Acquisition] = None):
         MetaDataBase.__init__(self, metadata, filePath, acquisitionDirectory=acquisitionDirectory)
         AnalysisManager.__init__(self, filePath)
-        self.fileFormat: ICMetaData.FileFormats = fileFormat
+        self.fileFormat: PwsMetaData.FileFormats = fileFormat
         self.dict['wavelengths'] = tuple(np.array(self.dict['wavelengths']).astype(float))
 
     def toDataClass(self, lock: mp.Lock = None) -> pwsdtd.PwsCube:
@@ -613,35 +613,35 @@ class ICMetaData(MetaDataBase, AnalysisManager):
         return self.dict['wavelengths']
 
     @classmethod
-    def loadAny(cls, directory, lock: mp.Lock = None, acquisitionDirectory: t_.Optional[Acquisition] = None) -> ICMetaData:
+    def loadAny(cls, directory, lock: mp.Lock = None, acquisitionDirectory: t_.Optional[Acquisition] = None) -> PwsMetaData:
         """
         Attempt to load from any file format.
 
         Args:
             directory: The file path to load the metadata from.
         Returns:
-            A new instance of `ICMetaData` loaded from file
+            A new instance of `PwsMetaData` loaded from file
         """
         try:
-            return ICMetaData.fromTiff(directory, lock=lock, acquisitionDirectory=acquisitionDirectory)
+            return PwsMetaData.fromTiff(directory, lock=lock, acquisitionDirectory=acquisitionDirectory)
         except:
             try:
-                return ICMetaData.fromOldPWS(directory, lock=lock, acquisitionDirectory=acquisitionDirectory)
+                return PwsMetaData.fromOldPWS(directory, lock=lock, acquisitionDirectory=acquisitionDirectory)
             except:
                 try:
-                    return ICMetaData.fromNano(directory, lock=lock, acquisitionDirectory=acquisitionDirectory)
+                    return PwsMetaData.fromNano(directory, lock=lock, acquisitionDirectory=acquisitionDirectory)
                 except:
                     raise OSError(f"Could not find a valid PWS image cube file at {directory}.")
 
     @classmethod
-    def fromOldPWS(cls, directory, lock: mp.Lock = None, acquisitionDirectory: t_.Optional[Acquisition] = None) -> ICMetaData:
+    def fromOldPWS(cls, directory, lock: mp.Lock = None, acquisitionDirectory: t_.Optional[Acquisition] = None) -> PwsMetaData:
         """
         Attempt to load from the old .mat file format.
 
         Args:
[…]
