### 2021-12-02 fc81608
minor plot improvements for ERworkflow
