### 2021-12-02 f48dbde
more minor plotting improvements ot ER workflow
