### 2021-12-03 943d5a5
Adding figures and more in-depth information.
