+++ paper/paper.md
@@ -25,18 +25,100 @@ Partial Wave Spectroscopic (PWS) Microscopy is a novel spectroscopic microscopy 
 detects the nanoscale mass density distribution within a biological sample. PWS microscopy achieves sensitivity to nanoscale structures within biological cells by using the spectroscopic content of microscope images, and quantitatively measures nanoarchitectural changes in cells associated with carcinogenesis[@Chandler:2016]. Additionally, PWS does not require exogenous labels and thus works even in unstained samples.
 
 `PWSpy` is a Python module dedicated to the analysis of PWS data. It includes a full suite of tools that prove useful in 
-analyzing experimental data. This includes object-oriented representations of the raw data involved as well as the analyzed output data and auxiliary data used such as ROIs, calibration files, and automated imaging metadata produced by the PWS Acquisition[@PWS_Acquisition] plugin for Micro-Manager[@Micro-Manager]. A single PWS measurement can be treated as a 3D image cube with dimensions (x, y, and wavenumber) and a single PWS experiment may consist of thousands of such images colocalized with widefield fluorescence, confocal microscopy, or other interferometric measurements. With `PWSpy`, it is trivial to skip the basics of preprocessing and get to the heart of extracting meaningful results from your experimental data. Basic operations such as normalization, camera defect compensation, and calibration can all be handled with the call of a single method so you can trust that they have been performed without error. Additionally, the library provides a means for conveniently loading and storing auxiliary data such as ROIs, notes, and analysis results. Utility functionality for generating visualizations, automatic colocalization, basic modeling of reflectance, loading an manipulation of Micro-Manager position list files, and more are provided in the `utility` subpackage.
+analyzing experimental data. This includes object-oriented representations of the raw data involved as well as the analyzed output data and auxiliary data used such as ROIs, calibration files, and automated imaging metadata produced by the PWS Acquisition[@PWS_Acquisition] plugin for Micro-Manager[@Micro-Manager]. A single PWS measurement can be treated as a 3D image cube with dimensions (x, y, and wavenumber) and a single PWS experiment may consist of thousands of such images colocalized with widefield fluorescence, confocal microscopy, or other interferometric image cubes. With `PWSpy`, it is trivial to skip the basics of preprocessing and get to the heart of extracting meaningful results from your experimental data. Basic operations such as normalization, camera non-linearity compensation, and calibration can all be handled with the call of a single method so you can trust that they have been performed without error. Additionally, the library provides a means for conveniently loading and storing auxiliary data such as ROIs, notes, and analysis results. Utility functionality for generating visualizations, automatic colocalization, basic modeling of reflectance, loading an manipulation of Micro-Manager position list files, and more are provided in the `utility` subpackage.
 
 
 # Statement of need
-Any analysis of raw data generated by a PWS experiment or other related interferometric imaging modalities[@Gladstein:2019] requires loading image data and metadata from a large variety of file formats and performing complex pre-processing steps before any of the real analysis begins. Up until now a great many single-use MATLAB scripts have been written to perform these tasks but a comprehensive and reusable library has never been developed.
-Any minor variations in how this pre-processing is performed can result in major differences in final analysis results and any script that fully implements all of the required processing becomes so long as to be unreadable. `PWSpy` provides an object-oriented interface for performing all common file I/O, pre-processing, and analysis tasks related to PWS. 
-This allows users to write succinct and readable scripts/software that can be trusted to process data correctly and which abstract away the variations in file formats between different PWS imaging systems. 
[…]
