### 2021-12-06 #3957 @jgostick
@nanthony21 the new paper looks much nicer.  The reviewers will drop in to begin playing with the software, but be patient.  This process could take a few months given the holiday shut-down will slow things.
