### 2022-01-03 c166729
update documentation of example
