+++ docs/source/examples.rst
@@ -47,3 +47,9 @@ Measuring Sigma using only a limited range of the OPD signal.
 .. literalinclude:: ../../examples/limitedOPDSigma.py
    :language: python
    :linenos:
+
+Generating new position lists to enable colocalized measurements on multiple systems.
+--------------------------------------------------------------
+.. literalinclude:: ../../examples/positionTransformation.py
+   :language: python
+   :linenos:
