### 2022-02-05 071ba50
Move PWSExperimentPath and PWSImagePath in __init__ to be individually defined in each example for better usability.
