+++ examples/__init__.py
@@ -33,10 +33,5 @@ This is a collection of example usages of the :mod:`pwspy` module.
     opdExampleScript
     positionTransformation
 """
-import pathlib as pl
 
-# Set this to a folder containing multiple "Cell{x}" acquisition folders. Some examples will use this.
-PWSExperimentPath = pl.Path(__file__).parent.parent / 'tests' / 'resources' / 'test_data' / 'sequencer'
 
-# Set this to the "Cell{X}" folder of a PWS acquisition. Most of the examples will then refer to this file.
-PWSImagePath = PWSExperimentPath / 'Cell1'
