+++ examples/compileResults.py
@@ -21,8 +21,12 @@ values within the ROIs for each acquisition. The compiled results are then place
 """
 import pandas
 from pwspy.analysis.compilation import PWSRoiCompiler, PWSCompilerSettings
-from examples import PWSExperimentPath
 import pwspy.dataTypes as pwsdt
+
+### User Variables ###
+PWSExperimentPath = ...  # Set this to a folder containing multiple "Cell{x}" acquisition folders. If you have downloaded the test dataset you can use `pl.Path(__file__).parent.parent / 'tests' / 'resources' / 'test_data' / 'sequencer'`
+######################
+
 
 tmpList = []
 compiler = PWSRoiCompiler(PWSCompilerSettings(reflectance=True, rms=True))
