### 2022-03-01 #3957 @pr4deepr
The software satisfies most of the criteria.
* I am using the test data to verify the functionality using examples in the docs. I have been communicating with @nanthony21 on the [PWSpy repo](https://github.com/BackmanLab/PWSpy/issues). I am not as familiar with Partial Wave Spectroscopic microscopy, but as far as I can tell, most functionality claims are validated. Just waiting on testing a few more.

* I am not sure about the Community contributions as I don't see `3) Seek support` within https://github.com/BackmanLab/PWSpy/blob/master/CONTRIBUTING.md . I may have misunderstood this point. @jgostick Should a contact be provided??
