### 2022-03-03 2db1113
remove old unused example files
