+++ examples/findOPDSurface/peakDetection/__init__.py
@@ -1,2 +1,0 @@
-"""This was an initial attempt that relied on detecting peaks in the spectra and then performing further processing
-to dtermine which peaks might be associated with reflections off the top surface of a cell."""
