+++ examples/findOPDSurface/peakDetection/findOPDSurface.py
@@ -1,212 +1,0 @@
-# Copyright 2018-2020 Nick Anthony, Backman Biophotonics Lab, Northwestern University
-#
-# This file is part of PWSpy.
-#
-# PWSpy is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
-#
-# PWSpy is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-# GNU General Public License for more details.
-#
-# You should have received a copy of the GNU General Public License
-# along with PWSpy.  If not, see <https://www.gnu.org/licenses/>.
-
-"""This script detects all local maxima in the spectra of a PWS image and then performs various morphological operations on the peak locations in an
- attempt to localize the cell surface. It is very finnick and needs to be adjusted for each individual image. I reccomend the "activeContour" script instead."""
-
-if __name__ == '__main__':
-    from pwspy.dataTypes import Acquisition
-    from glob import glob
-    import os
-    import scipy as sp
-    import numpy as np
-    from mpl_qt_viz.visualizers import PlotNd, MultiPlot
-    import matplotlib.pyplot as plt
-    import skimage.morphology as morph
-    import skimage.measure as meas
-    from examples.findOPDSurface.peakDetection.funcs import prune3dIm, Skel
-
-    # rootDir = r'G:\Data\NA_i_vs_NA_c\matchedNAi_largeNAc\cells'
-    rootDir = r'G:\Data\NA_i_vs_NA_c\smallNAi_largeNAc\cells'
-    anName = 'p0'
-    roiName = 'cell'
-    sigma = .5  # Blur radius in microns. greater sigma generally leads to greater sensitivity, with more contiguous regions of detected peaks. Also greater noise though.
-    opdCutoffLow = 3
-    opdCutoffHigh = 20
-
-    files = glob(os.path.join(rootDir, 'Cell*'))
-    acqs = [Acquisition(f) for f in files]
-
-    for acq in acqs:
-        try:
-            an = acq.pws.loadAnalysis(anName)
-        except OSError:
-            print("No analysis found. skipping")
-            continue
-
-        opd, opdIndex = an.reflectance.getOpd(useHannWindow=True)  # Using the Hann window should improve dynamic range and reduce false peaks from frequency leakage.
-
-        # Remove the low OPD signal.
-        idxLow = np.argmin(np.abs(opdIndex - opdCutoffLow))
-        idxHigh = np.argmin(np.abs(opdIndex - opdCutoffHigh))
-        opd = opd[:, :, idxLow:idxHigh]
-        opdIndex = opdIndex[idxLow:idxHigh]
-
-        # Blur laterally to denoise. This step is vital
-        Sigma = sigma / acq.pws.pixelSizeUm  # convert from microns to pixels
-        for i in range(opd.shape[2]):
-            opd[:, :, i] = sp.ndimage.filters.gaussian_filter(opd[:, :, i], Sigma, mode='reflect')
-
-        #Try to account for decreased sensitivity at higher opd
-        print("Adjusting OPD signal")
-        opd = opd * (opdIndex[None, None, :] ** 0.5)
-
-        print("Detect Peaks")
-        arr = np.zeros_like(opd, dtype=bool)
-        for i in range(opd.shape[0]):
-            for j in range(opd.shape[1]):
-                peaks, properties = sp.signal.find_peaks(opd[i, j, :])  # TODO vectorize? use kwargs?
[…]
