### 2022-03-03 908323f
rearrange example doc page. Minor fixes to various examples.
