### 2022-03-11 #3957 @pr4deepr
I've completed my review and am happy with the software and prompt responses by @nanthony21.
Please note that I am not as familiar with partial wave spectroscopy microscopy data and any functionality claims have been tested only using the examples and datasets provided in the library.

I think PWSpy is modular with an intuitive API, and makes data preprocessing and analysis straightforward.  This is a significant step up from single-use scripts which may only work for images from a specific instrument. The ability to save analysis settings and software versions in PWSpy is quite useful as this is often overlooked and underreported in many publications.

@nanthony21 has provided ample test data with example scripts and a detailed readme, so anyone can run this package.
