### 2022-03-21 #3957 @jgostick
Hi @nanthony21, having had no response from the other reviewer, I have removed them and just done the review myself.  Everything looks GREAT.  I was not able to test the 'functionality' thoroughly, since I don't really do this sort image analysis, but I do lots of other image analysis (tomograms of porous media) and I can clearly see that this package is first-rate.  So I will begin the process of accepting it!  Congrats and thanks for your patience.
