### 2022-03-21 #3957 @jgostick
Hi @nanthony21, can you address the items in first checklist below?

- [ ] Set article DOI with ``@editorialbot set <DOI here> as archive``
- [ ] Do a 'dry run' of acceptance with ``@editorialbot recommend accept``
