### 2022-03-25 #3957 @nanthony21
- [x] `Double check authors and affiliations (including ORCIDs)`
  I've checked the author, affiliations, and ORCIDs and everything is correct.
 - [x] `Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper.`
 I've created release 1.0.0 which includes the changes resulting from @pr4deepr's review.
 - [x]  `Archive the release on Zenodo/figshare/etc and post the DOI here.`
 The repository has been archived on Zenodo. DOI is : [DOI: 10.5281/zenodo.6381082](https://doi.org/10.5281/zenodo.6381082)
 - [x] `Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.`
