### 2022-03-31 #3957 @kyleniemeyer
Hi @nanthony21, I made a small PR that fixed some of the references in the paper, can you merge this? https://github.com/BackmanLab/PWSpy/pull/21

I will publish your article after that.
